test: keep incomplete bulk scan regressions together
